Fix asm tokens leaks because they used Vector and PVector (#5019)
* Fix asm tokens leaking horrendously because they used Vector and PVector.

* Fix some resource leaks related to asm tokens

* Fix some resource leaks of the hexagon module error cases.

* Fix possible NULL ptr dereference

* Fix leak of asm tokens by fun fini on RzAsmOp.

The RzAsmOp passed here could be a previously initialized one.
With allocated asm_tokens. Doing memset just leaks those tokens.

* Simplify the copy_asm_op function

* Always fini RzAsmOp before initializeing it.

* Set asm_toks to NULL after freeing

* Fix leak in asm_bf

* Fix leaks in test_token

* Initialize all RzAsmOps with 0 and free the asm_toks on init.

In practice rz_asm_op_init() is the best place to free the asm_toks reliably.
Simply because this is always called in the old code, before an instruction is disassembled.
Only in fini() is not enough, because there are too many place it is not called, leaking the tokens.

* Fresh mind, new attempt. Fini asmop at all occurrences of rz_asm_disassemble